added patreon link, just to see what response i get...
now pending is, once the chrome extension goes live, place the link in the link section where example.com is currently, and deploy this project. if the chrome webstore doesnt accept the extension then deploy this project first and then that, and then once chrome extension is live, come back and update this with that chrome link.